crypto_kx: add errors mod
