+++ crypto_kx/src/lib.rs
@@ -53,8 +53,10 @@
 mod keypair;
 mod keys;
 
+pub mod errors;
+
 pub use keypair::KeyPair;
 pub use keys::{
-    ClientSession as ClientSessionKeys, InvalidLength, Public as PublicKey, Secret as SecretKey,
+    ClientSession as ClientSessionKeys, Public as PublicKey, Secret as SecretKey,
     ServerSession as ServerSessionKeys, Session as SessionKey,
 };
